feat: added ingesting datasets table
